+++ CHANGELOG.md
@@ -12,11 +12,12 @@
  - Fixed cache leak in agent when deployments are canceled (#1883)
  - Improved robustness of modules update (#1885)
  - Removed environmental variables from agent report (#1891)
+ - Use asyncio subprocess instead of tornado subprocess (#1792)
 
 ## Added
  - Experimental data trace application (#1820, #1831)
  - Warning when shadowing variable (#1366)
- - Added support for compiler warnings (#1779)
+ - Added support for compiler warnings (#1779, #1905)
 
 # v 2020.1 (2020-02-19) Changes in this release: